proxy_api/requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
openai==0.27.8
python-dotenv==1.0.0
